Introduce DocumentViewHash, implement Hash for DocumentViewId
