+++ p2panda-rs/src/document/mod.rs
@@ -259,6 +259,7 @@
 mod document;
 mod document_id;
 mod document_view;
+mod document_view_hash;
 mod document_view_id;
 mod error;
 
@@ -267,5 +268,6 @@ use document::{build_graph, reduce};
 pub use document::{Document, DocumentBuilder};
 pub use document_id::DocumentId;
 pub use document_view::DocumentView;
+pub use document_view_hash::DocumentViewHash;
 pub use document_view_id::DocumentViewId;
 pub use error::{DocumentBuilderError, DocumentError, DocumentViewError};
